requirements.txt
adium-theme-ubuntu==0.3.4
autopep8==1.2.4
configobj==5.0.6
ecdsa==0.13
flake8==2.6.2
flake8-docstrings==0.2.8
mccabe==0.5.0
mysql-connector-python==2.0.4
mysql-utilities==1.6.1
paramiko==1.16.0
pep257==0.7.0
pep8==1.7.0
pexpect==4.0.1
ptyprocess==0.5
pycodestyle==2.0.0
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.2.3
pygobject==3.20.0
pyodbc==3.0.7
pysqlite==2.7.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.12.13
six==1.10.0
strictyaml==0.1.6
tails-installer==3.11.8
typing==3.5.2.2
unity-lens-photos==1.0
urlgrabber==3.9.1
vboxapi==1.0
virtualenv==15.0.2
